+++ CalligraphySample/src/main/java/uk/co/chrisjenx/calligraphy/sample/MainActivity.java
@@ -4,16 +4,21 @@ import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
-public class MainActivity extends ActionBarActivity {
+import static butterknife.ButterKnife.findById;
+
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        final Toolbar toolbar = findById(this, R.id.toolbar);
+        setSupportActionBar(toolbar);
 
         // Inject pragmatically
         getSupportFragmentManager()
@@ -21,13 +26,29 @@ public class MainActivity extends ActionBarActivity {
                 .replace(R.id.container, PlaceholderFragment.getInstance())
                 .commit();
 
-        new Handler(Looper.getMainLooper()).postDelayed(new Runnable() {
+
+        final Handler handler = new Handler(Looper.getMainLooper());
+        handler.postDelayed(new Runnable() {
             @Override
             public void run() {
-                setTitle("Calligraphy changed");
-                getSupportActionBar().setSubtitle("Added subtitle");
+//                toolbar.setTitle("Calligraphy Added");
+                toolbar.setSubtitle("Added subtitle");
             }
         }, 1000);
+
+        handler.postDelayed(new Runnable() {
+            @Override public void run() {
+                toolbar.setTitle(null);
+                toolbar.setSubtitle("Added subtitle");
+            }
+        }, 2000);
+
+        handler.postDelayed(new Runnable() {
+            @Override public void run() {
+                toolbar.setTitle("Calligraphy added back");
+                toolbar.setSubtitle("Added subtitle");
+            }
+        }, 3000);
     }
 
     /*