+++ calligraphy/src/main/java/uk/co/chrisjenx/calligraphy/CalligraphyConfig.java
@@ -2,12 +2,64 @@ package uk.co.chrisjenx.calligraphy;
 
 import android.os.Build;
 import android.text.TextUtils;
+import android.view.View;
+import android.widget.AutoCompleteTextView;
+import android.widget.Button;
+import android.widget.CheckBox;
+import android.widget.EditText;
+import android.widget.MultiAutoCompleteTextView;
+import android.widget.RadioButton;
+import android.widget.TextView;
+import android.widget.ToggleButton;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * Created by chris on 20/12/2013
  * Project: Calligraphy
  */
 public class CalligraphyConfig {
+
+    /**
+     * The default styles for the factory to lookup. The builder builds an extended immutable
+     * map of this with any additional custom styles.
+     */
+    private static final Map<Class<? extends TextView>, Integer> DEFAULT_STYLES = new HashMap<>();
+
+    static {
+        {
+            DEFAULT_STYLES.put(TextView.class, android.R.attr.textViewStyle);
+            DEFAULT_STYLES.put(Button.class, android.R.attr.buttonStyle);
+            DEFAULT_STYLES.put(EditText.class, android.R.attr.editTextStyle);
+            DEFAULT_STYLES.put(AutoCompleteTextView.class, android.R.attr.autoCompleteTextViewStyle);
+            DEFAULT_STYLES.put(MultiAutoCompleteTextView.class, android.R.attr.autoCompleteTextViewStyle);
+            DEFAULT_STYLES.put(CheckBox.class, android.R.attr.checkboxStyle);
+            DEFAULT_STYLES.put(RadioButton.class, android.R.attr.radioButtonStyle);
+            DEFAULT_STYLES.put(ToggleButton.class, android.R.attr.buttonStyleToggle);
+            if (CalligraphyUtils.canAddV7AppCompatViews()) {
+                addAppCompatViews();
+            }
+        }
+    }
+
+    /**
+     * AppCompat will inflate special versions of views for Material tinting etc,
+     * this adds those classes to the style lookup map
+     */
+    private static void addAppCompatViews() {
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatTextView.class, android.R.attr.textViewStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatButton.class, android.R.attr.buttonStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatEditText.class, android.R.attr.editTextStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatAutoCompleteTextView.class, android.R.attr.autoCompleteTextViewStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatMultiAutoCompleteTextView.class, android.R.attr.autoCompleteTextViewStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatCheckBox.class, android.R.attr.checkboxStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatRadioButton.class, android.R.attr.radioButtonStyle);
+        DEFAULT_STYLES.put(android.support.v7.widget.AppCompatCheckedTextView.class, android.R.attr.checkedTextViewStyle);
+    }
 
     private static CalligraphyConfig sInstance;
 
@@ -51,6 +103,19 @@ public class CalligraphyConfig {
      * Use Reflection to intercept CustomView inflation with the correct Context.
      */
     private final boolean mCustomViewCreation;
+    /**
+     * Use Reflection to try to set typeface for custom views if they has setTypeface method
+     */
+    private final boolean mCustomViewTypefaceSupport;
+    /**
+     * Class Styles. Build from DEFAULT_STYLES and the builder.
+     */
+    private final Map<Class<? extends TextView>, Integer> mClassStyleAttributeMap;
+    /**
+     * Collection of custom non-{@code TextView}'s registered for applying typeface during inflation
+     * @see uk.co.chrisjenx.calligraphy.CalligraphyConfig.Builder#addCustomViewWithSetTypeface(Class)
+     */
+    private final Set<Class<?>> hasTypefaceViews;
 
     protected CalligraphyConfig(Builder builder) {
         mIsFontSet = builder.isFontSet;
@@ -58,6 +123,11 @@ public class CalligraphyConfig {
         mAttrId = builder.attrId;
         mReflection = builder.reflection;
         mCustomViewCreation = builder.customViewCreation;
+        mCustomViewTypefaceSupport = builder.customViewTypefaceSupport;
+        final Map<Class<? extends TextView>, Integer> tempMap = new HashMap<>(DEFAULT_STYLES);
+        tempMap.putAll(builder.mStyleClassMap);
+        mClassStyleAttributeMap = Collections.unmodifiableMap(tempMap);
+        hasTypefaceViews = Collections.unmodifiableSet(builder.mHasTypefaceClasses);
     }
 
     /**
@@ -82,6 +152,18 @@ public class CalligraphyConfig {
         return mCustomViewCreation;
     }
 
+    public boolean isCustomViewTypefaceSupport() {
+        return mCustomViewTypefaceSupport;
+    }
+
+    public boolean isCustomViewHasTypeface(View view) {
+        return hasTypefaceViews.contains(view.getClass());
+    }
+
+    /* default */ Map<Class<? extends TextView>, Integer> getClassStyles() {
+        return mClassStyleAttributeMap;
+    }
+
     /**
      * @return the custom attrId to look for, -1 if not set.
      */
@@ -103,6 +185,10 @@ public class CalligraphyConfig {
          */
         private boolean customViewCreation = true;
         /**
+         * Use Reflection during view creation to try change typeface via setTypeface method if it exists
+         */
+        private boolean customViewTypefaceSupport = false;
+        /**
          * The fontAttrId to look up the font path from.
          */
         private int attrId = R.attr.fontPath;
@@ -114,6 +200,12 @@ public class CalligraphyConfig {
          * The default fontPath
          */
         private String fontAssetPath = null;
+        /**
+         * Additional Class Styles. Can be empty.
+         */
+        private Map<Class<? extends TextView>, Integer> mStyleClassMap = new HashMap<>();
+
+        private Set<Class<?>> mHasTypefaceClasses = new HashSet<>();
 
         /**
          * This defaults to R.attr.fontPath. So only override if you want to use your own attrId.
@@ -186,6 +278,37 @@ public class CalligraphyConfig {
          */
         public Builder disableCustomViewInflation() {
             this.customViewCreation = false;
+            return this;
+        }
+
+        /**
+         * Add a custom style to get looked up. If you use a custom class that has a parent style
+         * which is not part of the default android styles you will need to add it here.
+         *
+         * The Calligraphy inflater is unaware of custom styles in your custom classes. We use
+         * the class type to look up the style attribute in the theme resources.
+         *
+         * So if you had a {@code MyTextField.class} which looked up it's default style as
+         * {@code R.attr.textFieldStyle} you would add those here.
+         *
+         * {@code builder.addCustomStyle(MyTextField.class,R.attr.textFieldStyle}
+         *
+         * @param styleClass             the class that related to the parent styleResource. null is ignored.
+         * @param styleResourceAttribute e.g. {@code R.attr.textFieldStyle}, 0 is ignored.
+         * @return this builder.
+         */
+        public Builder addCustomStyle(final Class<? extends TextView> styleClass, final int styleResourceAttribute) {
+            if (styleClass == null || styleResourceAttribute == 0) return this;
+            mStyleClassMap.put(styleClass, styleResourceAttribute);
+            return this;
+        }
+
+        /**
+         * Register custom non-{@code TextView}'s which implement {@code setTypeface} so they can have the Typeface applied during inflation.
+         */
+        public Builder addCustomViewWithSetTypeface(Class<?> clazz) {
+            customViewTypefaceSupport = true;
+            mHasTypefaceClasses.add(clazz);
             return this;
         }