+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 #Changelog
 
+#2.2.0 (05/04/2016)
+- Added AppCompat Styles (AppCompatTextView will now pickup textViewStyle etc). Thanks @paul-turner
+- Fix for Toolbar not inflating `TextView`s upfront.
+
+#2.1.0 (27/04/2015)
+- Fixed #155, We now clone correctly.
+- Added Styles for Custom Views. (`builder.addCustomStyle(ToggleButton.class, android.R.attr.buttonStyleToggle)`)
+
 #2.0.2 (05/01/2015)
 - Fixed `CalligraphyConfig.Builder` missing return statements.
 - Fixed `createView()` getting the wrong parent context, Fixed: #135, #120